pyproject.toml
[project]
name = "test-mac-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.39.13",
    "ipykernel>=6.30.0",
    "pocketflow>=0.0.2",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "sentence-transformers>=5.0.0",
]
